+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2025_04_23_042347) do
+ActiveRecord::Schema[8.0].define(version: 2025_04_24_032754) do
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
@@ -219,7 +219,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_04_23_042347) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["assignment_id"], name: "index_selected_documents_on_assignment_id"
-    t.index ["google_doc_id"], name: "index_selected_documents_on_google_doc_id", unique: true
+    t.index ["google_doc_id"], name: "index_selected_documents_on_google_doc_id"
   end
 
   create_table "sessions", force: :cascade do |t|